requirements.txt
flask
twilio
python-dotenv
google-genai
gspread
oauth2client
pandas
